# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCRE API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/content-reports": {
      "get": {
        "tags": [
          "admin-content"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_api_v1_admin_content_reports_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(open|triaged|resolved|wont_fix)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportOut"
                  },
                  "title": "Response List Reports Api V1 Admin Content Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content-reports/{report_id}": {
      "patch": {
        "tags": [
          "admin-content"
        ],
        "summary": "Triage Report",
        "operationId": "triage_report_api_v1_admin_content_reports__report_id__patch",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/revisions": {
      "get": {
        "tags": [
          "admin-content"
        ],
        "summary": "List Revisions",
        "operationId": "list_revisions_api_v1_admin_content_revisions_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 12
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevisionRowOut"
                  },
                  "title": "Response List Revisions Api V1 Admin Content Revisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/content/revisions/{revision_id}": {
      "get": {
        "tags": [
          "admin-content"
        ],
        "summary": "Revision Detail",
        "operationId": "revision_detail_api_v1_admin_content_revisions__revision_id__get",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/revisions/{revision_id}/findings": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Add Finding",
        "operationId": "add_finding_api_v1_admin_content_revisions__revision_id__findings_post",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/findings/{finding_id}": {
      "patch": {
        "tags": [
          "admin-content"
        ],
        "summary": "Update Finding",
        "operationId": "update_finding_api_v1_admin_content_findings__finding_id__patch",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindingPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/revisions/{revision_id}/approve": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Approve",
        "operationId": "approve_api_v1_admin_content_revisions__revision_id__approve_post",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/revisions/{revision_id}/publish": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Publish",
        "operationId": "publish_api_v1_admin_content_revisions__revision_id__publish_post",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/units/{unit_id}/publish": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Publish Unit",
        "operationId": "publish_unit_api_v1_admin_content_units__unit_id__publish_post",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublishedOut"
                  },
                  "title": "Response Publish Unit Api V1 Admin Content Units  Unit Id  Publish Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/content/revisions/{revision_id}/withdraw": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Withdraw",
        "operationId": "withdraw_api_v1_admin_content_revisions__revision_id__withdraw_post",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApproveIn": {
        "properties": {
          "content_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Content Hash"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_hash"
        ],
        "title": "ApproveIn"
      },
...
      "FindingIn": {
        "properties": {
          "category": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Category"
          },
          "severity": {
            "type": "string",
            "enum": [
              "material",
              "minor"
            ],
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "severity",
          "description"
        ],
        "title": "FindingIn"
      },
      "FindingOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "category",
          "severity",
          "description",
          "status",
          "resolution_note",
          "created_at",
          "resolved_at"
        ],
        "title": "FindingOut"
      },
      "FindingPatch": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "resolved",
              "wont_fix"
            ],
            "title": "Status"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FindingPatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublishIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PublishIn"
      },
      "PublishedOut": {
        "properties": {
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "item_slug": {
            "type": "string",
            "title": "Item Slug"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "revision_id",
          "item_slug",
          "version"
        ],
        "title": "PublishedOut"
      },
...
      "ReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "item_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Title"
          },
          "activity_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Id"
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "page": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "triaged",
              "resolved",
              "wont_fix"
            ],
            "title": "Status"
          },
          "triage_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "revision_id",
          "item_title",
          "activity_id",
          "attempt_id",
          "category",
          "message",
          "page",
          "status",
          "triage_note",
          "created_at"
        ],
        "title": "ReportOut"
      },
      "ReportPatch": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "triaged",
              "resolved",
              "wont_fix"
            ],
            "title": "Status"
          },
          "triage_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ReportPatch"
      },
...
      "RevisionDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "item": {
            "additionalProperties": true,
            "type": "object",
            "title": "Item"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "file_status": {
            "type": "string",
            "title": "File Status"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "approved_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Hash"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          },
          "withdraw_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdraw Reason"
          },
          "source_path": {
            "type": "string",
            "title": "Source Path"
          },
          "source_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Commit"
          },
          "lint_errors": {
            "type": "integer",
            "title": "Lint Errors"
          },
          "lint_warnings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Lint Warnings"
          },
          "audio_pending": {
            "type": "boolean",
            "title": "Audio Pending"
          },
          "body": {
            "additionalProperties": true,
            "type": "object",
            "title": "Body"
          },
          "activities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Activities"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "type": "array",
            "title": "Findings"
          },
          "decisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Decisions"
          },
          "checklist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checklist"
          },
          "preview_practice": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preview Practice"
          },
          "preview_assessment": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preview Assessment"
          },
          "blockers": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Blockers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item",
          "version",
          "status",
          "file_status",
          "content_hash",
          "approved_hash",
          "approved_at",
          "published_at",
          "withdrawn_at",
          "withdraw_reason",
          "source_path",
          "source_commit",
          "lint_errors",
          "lint_warnings",
          "audio_pending",
          "body",
          "activities",
          "sources",
          "findings",
          "decisions",
          "checklist",
          "preview_practice",
          "preview_assessment",
          "blockers"
        ],
        "title": "RevisionDetailOut"
      },
      "RevisionRowOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "item_slug": {
            "type": "string",
            "title": "Item Slug"
          },
          "item_title": {
            "type": "string",
            "title": "Item Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "skill": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill"
          },
          "form_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Kind"
          },
          "unit": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "file_status": {
            "type": "string",
            "title": "File Status"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published"
          },
          "warnings": {
            "type": "integer",
            "title": "Warnings"
          },
          "audio_pending": {
            "type": "boolean",
            "title": "Audio Pending"
          },
          "open_material_findings": {
            "type": "integer",
            "title": "Open Material Findings"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "item_slug",
          "item_title",
          "kind",
          "skill",
          "form_kind",
          "unit",
          "version",
          "status",
          "file_status",
          "content_hash",
          "is_published",
          "warnings",
          "audio_pending",
          "open_material_findings",
          "created_at"
        ],
        "title": "RevisionRowOut"
      },
...
      "WithdrawIn": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "WithdrawIn"
      }
    }
  }
}